refactor(agent-core): LlmError を llm-error.ts に分離し llm-client の lazy singleton を廃止 #132
agent.ts が LlmError のためだけに llm-client.ts をインポートしていたため、
agent.test.ts のロード時に Anthropic クライアントが生成されて llm-client.test.ts の
mock.module が効かなくなっていた。

LlmError を llm-error.ts に切り出すことで依存グラフを整理する。
- agent.ts → llm-error.ts（Anthropic SDK なし）
- llm-client.ts → llm-error.ts（LlmError を再利用）
- llm-client.ts のモジュールロード時初期化（fail-fast）を復元
- llm-error.ts から LlmError を export するよう index.ts を更新
