+++ packages/agent-core/src/index.ts
@@ -7,4 +7,5 @@ export type {
 export { runExecution } from "./engine.ts";
 export type { AgentEvent } from "./events.ts";
 export type { LlmInput } from "./llm-client.ts";
-export { LlmError, streamAgentMessage } from "./llm-client.ts";
+export { streamAgentMessage } from "./llm-client.ts";
+export { LlmError } from "./llm-error.ts";
