+++ packages/agent-core/src/llm-client.ts
@@ -1,14 +1,12 @@
 /**
  * Anthropic SDK の薄いラッパー。
  *
- * クライアントは lazy singleton で初回呼び出し時に初期化する。
- * モジュールロード時に副作用がないため、fake stream を DI するテストで
- * `LLM_API_KEY` 不要かつ `mock.module` が正しく適用される。
+ * モジュールロード時に `LLM_API_KEY` を検証する副作用がある。
  * エラーは `LlmError` に統一し、AbortSignal による中断はそのまま伝播する。
  */
 
-import type { AgentFailReason } from "@agent-team-studio/shared";
 import Anthropic from "@anthropic-ai/sdk";
+import { LlmError } from "./llm-error.ts";
 
 /** LLM 呼び出しのパラメータ（モデル・プロンプト・温度・トークン上限）。 */
 export type LlmInput = {
@@ -19,39 +17,22 @@ export type LlmInput = {
   max_tokens: number;
 };
 
-/** LLM API 失敗を表すカスタムエラー。`failReason` で失敗の種別を識別する。 */
-export class LlmError extends Error {
-  readonly failReason: AgentFailReason;
-
-  constructor(
-    failReason: AgentFailReason,
-    message: string,
-    options?: ErrorOptions,
-  ) {
-    super(message, options);
-    this.name = "LlmError";
-    this.failReason = failReason;
-  }
+const apiKey = process.env.LLM_API_KEY;
+if (!apiKey) {
+  throw new Error("LLM_API_KEY is not set");
 }
 
-let _client: Anthropic | null = null;
+const clientOptions: ConstructorParameters<typeof Anthropic>[0] = {
+  apiKey,
+  maxRetries: 3,
+  timeout: 120_000,
+};
 
-/** 初回呼び出し時にクライアントを生成してキャッシュする。 */
-function getClient(): Anthropic {
-  if (_client) return _client;
-  const apiKey = process.env.LLM_API_KEY;
-  if (!apiKey) throw new Error("LLM_API_KEY is not set");
-  const options: ConstructorParameters<typeof Anthropic>[0] = {
-    apiKey,
-    maxRetries: 3,
-    timeout: 120_000,
-  };
-  if (process.env.LLM_BASE_URL) {
-    options.baseURL = process.env.LLM_BASE_URL;
-  }
-  _client = new Anthropic(options);
-  return _client;
+if (process.env.LLM_BASE_URL) {
+  clientOptions.baseURL = process.env.LLM_BASE_URL;
 }
+
+const client = new Anthropic(clientOptions);
 
 /**
  * LLM からのテキストを非同期ストリームで返す。
@@ -63,7 +44,6 @@ export async function* streamAgentMessage(
   input: LlmInput,
   signal?: AbortSignal,
 ): AsyncIterable<string> {
-  const client = getClient();
   try {
     const stream = client.messages.stream(
       {
